# apis/swagger.yml
...
swagger: "2.0"
schemes:
  - "http"
  - "https"
produces:
  - "application/json"
  - "text/plain"
consumes:
  - "application/json"
  - "text/plain"
basePath: "/v1.24"
info:
  title: "Pouch Engine API"
  version: "1.24"
  description: |
    API is an HTTP API served by Pouch Engine.
paths:
...
  /images/{imageid}/history:
    get:
      summary: "Get an image's history"
      description: "Return the history of each layer of image"
      operationId: "ImageHistory"
      produces:
        - "application/json"
      responses:
        200:
          description: "no error"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/HistoryResultItem"
        404:
          $ref: "#/responses/404ErrorResponse"
        500:
          $ref: "#/responses/500ErrorResponse"
      parameters:
        - $ref: "#/parameters/imageid"
  /images/json:
    get:
      summary: "List Images"
      description: "Return a list of stored images."
      operationId: "ImageList"
      produces:
        - "application/json"
      responses:
        200:
          description: "Summary image data for the images matching the query"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ImageInfo"
          examples:
            application/json:
              - Id: "sha256:e216a057b1cb1efc11f8a268f37ef62083e70b1b38323ba252e25ac88904a7e8"
                ParentId: ""
                RepoTags:
                  - "ubuntu:12.04"
                  - "ubuntu:precise"
                RepoDigests:
                  - "ubuntu@sha256:992069aee4016783df6345315302fa59681aae51a8eeb2f889dea59290f21787"
                Created: 1474925151
                Size: 103579269
                VirtualSize: 103579269
                SharedSize: 0
                Labels: {}
                Containers: 2
              - Id: "sha256:3e314f95dcace0f5e4fd37b10862fe8398e3c60ed36600bc0ca5fda78b087175"
                ParentId: ""
                RepoTags:
                  - "ubuntu:12.10"
                  - "ubuntu:quantal"
                RepoDigests:
                  - "ubuntu@sha256:002fba3e3255af10be97ea26e476692a7ebed0bb074a9ab960b2e7a1526b15d7"
                  - "ubuntu@sha256:68ea0200f0b90df725d99d823905b04cf844f6039ef60c60bf3e019915017bd3"
                Created: 1403128455
                Size: 172064416
                VirtualSize: 172064416
                SharedSize: 0
                Labels: {}
                Containers: 5
        500:
          $ref: "#/responses/500ErrorResponse"
      parameters:
        - name: "all"
          in: "query"
          description: "Show all images. Only images from a final layer (no children) are shown by default."
          type: "boolean"
        - name: "filters"
          in: "query"
          description: |
            A JSON encoded value of the filters (a `map[string][]string`) to process on the images list. Available filters:

            - `before`=(`<image-name>[:<tag>]`,  `<image id>` or `<image@digest>`)
            - `reference`=(`<image-name>[:<tag>]`)
            - `since`=(`<image-name>[:<tag>]`,  `<image id>` or `<image@digest>`)
          type: "string"
        - name: "digests"
          in: "query"
          description: "Show digest information as a `RepoDigests` field on each image."
          type: "boolean"
...
definitions:
  Error:
    type: "object"
    properties:
      message:
        type: string
...
  ContainerConfig:
    type: "object"
    description: "Configuration for a container that is portable between hosts"
    required: 
      - Image
    properties:
      Hostname:
        description: "The hostname to use for the container, as a valid RFC 1123 hostname."
        type: "string"
        format: hostname
        minLength: 1
      Domainname:
        description: "The domain name to use for the container."
        type: "string"
      User:
        description: "The user that commands are run as inside the container."
        type: "string"
      AttachStdin:
        description: "Whether to attach to `stdin`."
        type: "boolean"
        x-nullable: false
      AttachStdout:
        description: "Whether to attach to `stdout`."
        type: "boolean"
        x-nullable: false
        default: true
      AttachStderr:
        description: "Whether to attach to `stderr`."
        type: "boolean"
        x-nullable: false
        default: true
      DisableNetworkFiles:
        description: "Whether to generate the network files(/etc/hostname, /etc/hosts and /etc/resolv.conf) for container."
        type: "boolean"
        x-nullable: false
        default: false
      ExposedPorts:
        description: "An object mapping ports to an empty object in the form:`{<port>/<tcp|udp>: {}}`"
        type: "object"
        additionalProperties:
          type: "object"
          enum:
            - {}
          default: {}
      Tty:
        description: "Attach standard streams to a TTY, including `stdin` if it is not closed."
        type: "boolean"
        x-nullable: false
      OpenStdin:
        description: "Open `stdin`"
        type: "boolean"
        x-nullable: false
      StdinOnce:
        description: "Close `stdin` after one attached client disconnects"
        type: "boolean"
        x-nullable: false
      Env:
        description: |
          A list of environment variables to set inside the container in the form `["VAR=value", ...]`. 
          A variable like "A=" means setting env A in container to be empty value.
          And a variable without `=` is removed from the environment, rather than to have an empty value.
        type: "array"
        items:
          type: "string"
      Cmd:
        description: "Command to run specified an array of strings."
        type: "array"
        items:
          type: "string"
      ArgsEscaped:
        description: "Command is already escaped (Windows only)"
        type: "boolean"
        x-nullable: false
      Image:
        description: "The name of the image to use when creating the container"
        type: "string"
        x-nullable: false
      Volumes:
        description: "An object mapping mount point paths inside the container to empty objects."
        type: "object"
        additionalProperties:
          type: "object"
          enum:
            - {}
          default: {}
      WorkingDir:
        description: "The working directory for commands to run in."
        type: "string"
      Entrypoint:
        description: |
          The entry point for the container as a string or an array of strings.
          If the array consists of exactly one empty string (`[""]`) then the entry point is reset to system default.
        type: "array"
        items:
          type: "string"
      NetworkDisabled:
        description: "Disable networking for the container."
        type: "boolean"
      MacAddress:
        description: "MAC address of the container."
        type: "string"
      OnBuild:
        description: "`ONBUILD` metadata that were defined."
        type: "array"
        items:
          type: "string"
      Labels:
        description: "User-defined key/value metadata."
        type: "object"
        additionalProperties:
          type: "string"
      StopSignal:
        description: "Signal to stop a container as a string or unsigned integer."
        type: "string"
        default: "SIGTERM"
        x-nullable: false
      StopTimeout:
        description: "Timeout to stop a container in seconds."
        type: "integer"
        minimum: 0
        default: 10
      Shell:
        description: "Shell for when `RUN`, `CMD`, and `ENTRYPOINT` uses a shell."
        type: "array"
        items:
          type: "string"
      Rich:
        type: "boolean"
        description: "Whether to start container in rich container mode. (default false)"
        x-nullable: false
      RichMode:
        type: "string"
        description: "Choose one rich container mode.(default dumb-init)"
        enum:
          - "dumb-init"
          - "sbin-init"
          - "systemd"
      InitScript:
        type: "string"
        description: "Initial script executed in container. The script will be executed before entrypoint or command"
      DiskQuota:
        type: "object"
        description: |
          Set disk quota for container.
          Key is the dir in container.
          Value is disk quota size for the dir.
          / means rootfs dir in container.
          .* includes rootfs dir and all volume dir.
        x-nullable: true
        additionalProperties:
          type: "string"
      SpecAnnotation:
        description: "annotations send to runtime spec."
        type: "object"
        additionalProperties:
          type: "string"
      QuotaID:
        type: "string"
        description: |
          Set disk quota by specified quota id. 
          If QuotaID <= 0, it means pouchd should allocate a unique quota id by sequence automatically.
          By default, a quota ID is mapped to only one container. And one quota ID can include several mountpoint.
      NetPriority:
        description: "net priority."
        type: "integer"
        default: 0
      SpecificID:
        type: "string"
        description: |
          Create container with given id.
          MinLength: 64
          MaxLength: 64
          The characters of given id should be in 0123456789abcdef.
          By default, given id is unnecessary.
      Snapshotter:
        description: |
            The snapshotter container choose, can be different with
            default snapshotter. The Field only set through hook plugin.
        type: "string"
...
  ImageInfo:
    description: "An object containing all details of an image at API side"
    type: "object"
    properties:
      Id:
        description: "ID of an image."
        type: "string"
        x-nullable: false
      RepoTags:
        description: "repository with tag."
        type: "array"
        items:
          type: "string"
      RepoDigests:
        description: "repository with digest."
        type: "array"
        items:
          type: "string"
      CreatedAt:
        description: "time of image creation."
        type: "string"
        x-nullable: false
      Size:
        description: "size of image's taking disk space."
        type: "integer"
        x-nullable: false
      Config:
        $ref: "#/definitions/ContainerConfig"
      Architecture:
        description: "the CPU architecture."
        type: "string"
        x-nullable: false
      Os:
        description: "the name of the operating system."
        type: "string"
        x-nullable: false
      RootFS:
        description: "the rootfs key references the layer content addresses used by the image."
        type: "object"
        required: [Type]
        properties:
          Type:
            description: "type of the rootfs"
            type: "string"
            x-nullable: false
          Layers:
            description: "an array of layer content hashes"
            type: "array"
            items:
              type: "string"
          BaseLayer:
            description: "the base layer content hash."
            type: "string"
...
  HistoryResultItem:
    description: "An object containing image history at API side."
    type: "object"
    required: [ID, Created, CreatedBy, Author, Comment, EmptyLayer, Size]
    properties:
      ID:
        description: "ID of each layer image."
        type: "string"
        x-nullable: false
      Created:
        description: "the combined date and time at which the layer was created."
        type: "integer"
        format: "int64"
        x-nullable: false
      CreatedBy:
        description: "the command which created the layer."
        type: "string"
        x-nullable: false
      Author:
        description: "the author of the build point."
        type: "string"
        x-nullable: false
      Comment:
        description: "a custom message set when creating the layer."
        type: "string"
        x-nullable: false
      EmptyLayer:
        description: "mark whether the history item created a filesystem diff or not."
        type: "boolean"
        x-nullable: false
      Size:
        description: "size of each layer image."
        type: "integer"
        format: "int64"
        x-nullable: false
...
parameters:
...
  imageid:
    name: imageid
    in: path
    required: true
    description: Image name or id
    type: string
responses:
...
  404ErrorResponse:
    description: An unexpected 404 error occurred.
    schema:
      $ref: "#/definitions/Error"
  500ErrorResponse:
    description: An unexpected server error occurred.
    schema:
      $ref: "#/definitions/Error"